+++ src/Resources/app/administration/src/module/heptacom-admin-open-auth-client/page/edit/index.js
@@ -1,3 +1,4 @@
+import './heptacom-admin-open-auth-client-edit-page.scss';
 import template from './heptacom-admin-open-auth-client-edit-page.html.twig';
 
 const { Component, Context, Mixin, Data } = Shopware;